Prevent force insertion of duplicate revisions
Only force insert revisions which are not already in the database,
preventing multiple roots with the same document ID and revision
for being created.

When it is attempted to force insert a revision which is already
present in the database, a `DocumentException` is thrown. Which contains
information about the revision and doc ID that was going to be inserted
into the database.

Fixes #329